requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
langchain-groq
langchain-google-genai
langchain-chroma
chromadb
sentence-transformers
pandas
numpy
scikit-learn
streamlit
python-dotenv
openai
groq
fastapi
uvicorn[standard]
python-multipart
langchain-text-splitters
pypdf
